requirements.txt
# Web Framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Settings
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
greenlet>=3.1.0

# Supabase
supabase>=2.4.0

# LangGraph / Agent
langgraph>=0.1.0
langchain-core>=0.2.0

# HTTP Client
httpx>=0.27.0

# Data Analysis
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.13.0
openpyxl>=3.1.0

# Vector DB
pinecone>=3.2.0

# Security (sandbox)
RestrictedPython>=7.1

# Utilities
python-dotenv>=1.0.0